flask_backend/requirements-ci.txt
sqlalchemy
flask-sqlalchemy
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
flask-jwt-extended
flask-cors
